Revert "Feature/adapters"
