+++ .idea/jarRepositories.xml
@@ -26,10 +26,5 @@
       <option name="name" value="Google" />
       <option name="url" value="https://dl.google.com/dl/android/maven2/" />
     </remote-repository>
-    <remote-repository>
-      <option name="id" value="maven" />
-      <option name="name" value="maven" />
-      <option name="url" value="https://oss.jfrog.org/libs-snapshot" />
-    </remote-repository>
   </component>
 </project>
