+++ build.gradle
@@ -19,7 +19,6 @@ allprojects {
         google()
         mavenCentral()
         jcenter() // Warning: this repository is going to shut down soon
-        maven { url "https://oss.jfrog.org/libs-snapshot" }
     }
 }
